requirements.txt
streamlit>=1.33.0

langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-tavily>=0.1.0

huggingface_hub>=0.23.0
pydantic>=2.7.0

python-dotenv>=1.0.0